renames utils to scrape-utils
